Move integration tests of Spine Web to io.spine.web.test package.
