+++ integration-tests/web-tests/src/main/java/io/spine/web/test/given/Application.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.web.test;
+package io.spine.web.test.given;
 
 import io.spine.server.BoundedContext;
 import io.spine.server.CommandService;
