+++ integration-tests/web-tests/src/main/java/io/spine/web/test/given/TaskRepository.java
@@ -18,32 +18,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.web.test;
+package io.spine.web.test.given;
 
-import io.spine.server.aggregate.Aggregate;
-import io.spine.server.aggregate.Apply;
-import io.spine.server.command.Assign;
+import io.spine.server.aggregate.AggregateRepository;
 
 /**
- * An aggregate that is used to create projects.
+ * A repository for the task aggregates.
  *
- * @author Mykhailo Drachuk
+ * @author Dmytro Dashenkov
  */
-public class ProjectAggregate extends Aggregate<ProjectId, Project, ProjectVBuilder> {
-
-    public ProjectAggregate(ProjectId id) {
-        super(id);
-    }
-
-    @Assign
-    ProjectCreated handle(CreateProject command) {
-        return ProjectCreatedVBuilder.newBuilder()
-                                     .setId(command.getId())
-                                     .build();
-    }
-
-    @Apply
-    private void on(ProjectCreated event) {
-        getBuilder().setId(event.getId());
-    }
+class TaskRepository extends AggregateRepository<TaskId, TaskAggregate> {
 }
