+++ integration-tests/web-tests/src/main/java/io/spine/web/test/given/TestQueryServlet.java
@@ -18,36 +18,26 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.web.test;
+package io.spine.web.test.given;
 
-import io.spine.server.aggregate.Aggregate;
-import io.spine.server.aggregate.Apply;
-import io.spine.server.command.Assign;
+import io.spine.web.firebase.FirebaseQueryBridge;
+import io.spine.web.firebase.FirebaseQueryServlet;
+
+import javax.servlet.annotation.WebServlet;
 
 /**
- * An aggregate with the state of type {@code spine.web.test.Task}.
+ * The query side endpoint of the application.
  *
  * @author Dmytro Dashenkov
  */
-public class TaskAggregate extends Aggregate<TaskId, Task, TaskVBuilder> {
+@WebServlet("/query")
+@SuppressWarnings("serial")
+public class TestQueryServlet extends FirebaseQueryServlet {
 
-    public TaskAggregate(TaskId id) {
-        super(id);
-    }
-
-    @Assign
-    TaskCreated handle(CreateTask command) {
-        return TaskCreatedVBuilder.newBuilder()
-                                  .setId(command.getId())
-                                  .setName(command.getName())
-                                  .setDescription(command.getDescription())
-                                  .build();
-    }
-
-    @Apply
-    private void on(TaskCreated event) {
-        getBuilder().setId(event.getId())
-                    .setName(event.getName())
-                    .setDescription(event.getDescription());
+    public TestQueryServlet() {
+        super(FirebaseQueryBridge.newBuilder()
+                                 .setQueryService(Server.application().getQueryService())
+                                 .setDatabase(FirebaseClient.database())
+                                 .build());
     }
 }
