+++ integration-tests/web-tests/src/main/proto/spine/web/test/given/events.proto
@@ -1,19 +1,19 @@
 syntax = "proto3";
 
-package spine.web.test;
+package spine.web.test.given;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.web.test";
-option java_outer_classname = "CommandsProto";
+option java_package = "io.spine.web.test.given";
+option java_outer_classname = "EventsProto";
 option java_multiple_files = true;
 option java_generate_equals_and_hash = true;
 
-import "spine/web/test/task.proto";
-import "spine/web/test/project.proto";
+import "spine/web/test/given/task.proto";
+import "spine/web/test/given/project.proto";
 
-message CreateTask {
+message TaskCreated {
 
     TaskId id = 1;
 
@@ -22,7 +22,7 @@ message CreateTask {
     string description = 3;
 }
 
-message CreateProject {
-
+message ProjectCreated {
+    
     ProjectId id = 1;
 }
