+++ integration-tests/web-tests/src/main/proto/spine/web/test/given/project.proto
@@ -1,11 +1,11 @@
 syntax = "proto3";
 
-package spine.web.test;
+package spine.web.test.given;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.web.test";
+option java_package = "io.spine.web.test.given";
 option java_outer_classname = "ProjectProto";
 option java_multiple_files = true;
 option java_generate_equals_and_hash = true;
